Move displayName() to unmanaged objects & fixed all unit tests
