+++ Rocket.Chat/Models/User/User.swift
@@ -55,6 +55,32 @@ extension User: UnmanagedConvertible {
     }
 }
 
+// MARK: Display Name
+
+extension User {
+
+    func displayName() -> String {
+        guard let validatedUser = validated() else {
+            return ""
+        }
+
+        let username = validatedUser.username ?? ""
+
+        guard let settings = AuthSettingsManager.settings else {
+            return username
+        }
+
+        if let name = validatedUser.name {
+            if settings.useUserRealName && !name.isEmpty {
+                return name
+            }
+        }
+
+        return username
+    }
+
+}
+
 // MARK: Avatar URL
 
 extension User {
